#111 [feat] 통계화면 기능 일부 구현 1. 닉네임 연결 2. 리뷰 총 평균 점수 3. 기대평 작성 총 갯수
