+++ app/src/main/res/layout/fragment_detail_location.xml
@@ -14,7 +14,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="24dp"
-            android:layout_marginTop="30dp"
+            android:layout_marginTop="30dp"          
             android:text="@string/map"
             android:textColor="@color/text_color"
             android:textSize="25sp"
